refactor(utilities): continued reorg/ cleanup of utilities
